rename mining-tools.lua to player-mining-tools.lua
there's bad naming here, but I want mining.lua for stuff to do with real resource mining
