+++ control.lua
@@ -13,7 +13,7 @@ local fa_tutorial = require("scripts.tutorial-system")
 local fa_sectors = require("scripts.building-vehicle-sectors")
 local fa_menu_search = require("scripts.menu-search")
 local fa_building_tools = require("scripts.building-tools")
-local fa_mining_tools = require("scripts.mining-tools")
+local fa_mining_tools = require("scripts.player-mining-tools")
 local fa_rails = require("scripts.rails")
 local fa_rail_builder = require("scripts.rail-builder")
 local fa_trains = require("scripts.trains")
